feat: Allow user to specify envrionment variables for scripts
This allows the user to specify env vars for a script. It allows to specify a set of env vars for
multiple scripts at once. So, for examplel, specifying env: { test: { foo: 'bar' } } in the config
will run 'test.watch', 'test.mystuff' 'test.otherstuff' will foo=bar as an env var.