fix(default): Now default scripts will be a fallback only
**What**:

This adds logic to check first for non-default scripts, then fallback on
default scripts if none are found that satisfies the input

**Why**:

Because #25

**How**:

Create a clone of the config object and remove all default scripts,
check the prefix with that, if none is found attempt again with the
original config object.

Closes #25

BREAKING CHANGE: You're unlikely to have this problem, but if you had a
script that matched the prefix `def` for example, the default script
would be run. Now, your specific script will be run instead.