fix(cli): improve handling of forwarding args/flags
**What**: This migrates us from `commander` and `omelette` to `yargs`
and in the process changes how we handle forwarding of arguments and
flags. We no longer just arbitrarily forward arguments to every script.
Also instead of comma separated scripts, they're simply separated by a
space. If you want to pass args/flags to a script, then you put it in
quotes.

**Why**: Makes this more predicatble. Also `yargs` is more powerful in
general anyway.

**How**: Magic... Will comment inline...